fix gitignore: remove DerivedData
