--- InvestmentApp/InvestmentApp/Utils/Extensions/UITableViewCell+reuseIdentifier.swift
@@ -1,21 +1,0 @@
-//
-//  UITableViewCell+reuseIdentifier.swift
-//  InvestmentApp
-//
-//  Created by Jessica Lewinter on 09/06/20.
-//  Copyright © 2020 Jessica Lewinter. All rights reserved.
-//
-
-import UIKit
-
-protocol Identifiable {
-    static var reuseIdentifier: String { get }
-}
-
-extension Identifiable {
-    static var reuseIdentifier: String {
-        String(describing: Self.self)
-    }
-}
-
-extension UITableViewCell: Identifiable {}
